refactor(profiles): decompose profile.rs into focused modules [OPE-384]
Split the monolithic 745-line profile.rs into a multi-file module:
- profile/types.rs (~120 lines): ExtensionRef, SubRecipeRef, ParameterRef, ProviderFallback, ProfileSettings
- profile/mod.rs (~100 lines): AgentProfile struct, impl, and YamlDefinition trait
- profile/tests.rs (~480 lines): all existing tests preserved
